requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
dj-database-url==3.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
gunicorn==23.0.0
Pillow==10.4.0
